floodsafe-server/requirements.txt
fastapi[all]
pytorch
torchvision
selenium
pandas
matplotlib
scikit-learn
pillow